requirements.txt
streamlit==0.94.0
pillow==8.3.1
python-dotenv==0.24.0
pandas==1.3.3
langchain==0.1.2  # Assuming this is the correct version of langchain
openai==0.27.0
pandasai==0.1.0  # Assuming this is the correct version of pandasai
